Change the oscar capabilities variable to be a guint64 everywhere instead of an enum, which can be 32 bits
--HG--
extra : convert_revision : 8df884656e9b2e225d429b9121f8449894019d96

+++ libpurple/protocols/oscar/tlv.c
@@ -407,7 +407,7 @@ int aim_tlvlist_add_str(GSList **list, const guint16 type, const char *value)
  * @param caps Bitfield of capability flags to send
  * @return The size of the value added.
  */
-int aim_tlvlist_add_caps(GSList **list, const guint16 type, const guint32 caps, const char *mood)
+int aim_tlvlist_add_caps(GSList **list, const guint16 type, const guint64 caps, const char *mood)
 {
 	guint8 buf[256]; /* TODO: Don't use a fixed length buffer */
 	ByteStream bs;
